Connect button with auth
